requirements.txt
absl-py==1.4.0
algoliasearch==3.0.0
anyio==3.6.2
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
binwalk==2.2.0
bleach==6.0.0
blinker==1.4
boxes==0.0.0
Brlapi==0.7.0
cachetools==5.2.0
cairocffi==1.4.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
chrome-gnome-shell==0.0.0
click==8.1.2
colorama==0.4.3
colorclass==2.2.2
comm==0.1.3
command-not-found==0.3
contextlib2==21.6.0
cryptography==2.8
cssselect==1.1.0
cssutils==2.4.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.6.6
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
demjson==2.2.4
diff-match-patch==20200713
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.2
django-cors-headers==3.2.1
django-import-export==2.0.1
django-rest-passwordreset==1.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
execnet==1.9.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
Flask==2.1.1
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.8.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.3
importlib-resources==5.12.0
iniconfig==1.1.1
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.1
jsonschema==4.17.3
jtutils==0.0.8
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
keras==2.12.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
leven==1.0.4
libclang==16.0.0
lolcat==1.4
louis==3.12.0
lxml==4.9.0
macaroonbakery==1.3.1
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.0.4
mock==4.0.3
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.10.4
nose==1.3.7
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-contrib-python==4.5.5.64
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.1.0
packaging==20.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
path==16.4.0
path.py==12.5.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pip-upgrader==1.4.15
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
premailer==3.10.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.5.1
psycopg2==2.8.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.9.1
pycparser==2.21
pycups==1.9.73
pydbus==0.6.0
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==22.0.0
pyparsing==2.4.6
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
pyte==0.8.0
pytest==7.1.2
pytest-shutil==1.7.0
python-apt==2.0.1+ubuntu0.20.4.1
python-csv==0.0.13
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-json-logger==2.0.7
python-Levenshtein==0.12.0
python-magic==0.4.16
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
reportlab==3.5.34
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scipy==1.10.1
SecretStorage==2.3.1
selenium==4.2.0
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
ssh-import-id==5.10
stack-data==0.6.2
systemd-python==234
tablib==0.14.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
terminado==0.17.1
terminaltables==3.1.10
thefuck==3.32
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
ulauncher==5.15.2
unattended-upgrades==0.1
urllib3==1.26.9
virtualenv==20.17.1
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.1.1
whitenoise==6.3.0
widgetsnbextension==4.0.6
wrapt==1.14.1
wsproto==1.1.0
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
yagmail==0.15.277
zipp==3.7.0
